+++ packages/vx-demo/src/docs-v2/examples/vx-linktypes/useForceUpdate.ts
@@ -1,0 +1,6 @@
+import { useState } from 'react';
+
+export default function useForceUpdate() {
+  const [, setValue] = useState<number>(0);
+  return () => setValue(value => value + 1); // update state to force render
+}